Extracted an Asteroid base class and specs to a shared_example
