--- spec/lib/large_asteroid_spec.rb
@@ -1,44 +1,0 @@
-require "large_asteroid"
-
-RSpec.describe LargeAsteroid do
-  describe ".new" do
-    it { is_expected.to be_an_instance_of LargeAsteroid }
-
-    it "has an image" do
-      expect(subject.image).to be_an_instance_of Gosu::Image
-    end
-
-    it "has a shape" do
-      expect(subject.shape).to be_a CP::Shape
-    end
-
-    it "collides as an asteroid" do
-      expect(subject.shape.collision_type).to be :asteroid
-    end
-
-    it "associates itself to its shape" do
-      expect(subject.shape.object).to be subject
-    end
-
-    it "has a body" do
-      expect(subject.body).to be_an_instance_of CP::Body
-    end
-  end
-
-  describe "#points" do
-    it "is worth 20 points" do
-      expect(subject.points).to be 20
-    end
-  end
-
-  describe "chunks" do
-    it "breaks into 2 Medium asteroids" do
-      chunks = subject.chunks
-
-      expect(chunks.length).to be 2
-      chunks.each do |chunk|
-        expect(chunk).to be_an_instance_of MediumAsteroid
-      end
-    end
-  end
-end
